Fix memory leak bugs
In this commit, I fix several memory leak bugs. Most of them are just
forgeting free the allocated memory. There is a special case in
timeConverter function class. Previously, the convers is a member
declared as static member in a class. The variable is a vector
attempting to containing several time_converter_base_t pointers. Those
pointers are initialized in a .cpp file. The pointers cannot be deleted
in a destructor like ordinary way does, since the destructor will be
invoked when an object is deleted, the static variable would be used
across all objects in same types. In view of this, I wrap the pointers
in another class in ordinary containers and remain the converters in
timeConverter class static. This solves the problem.